feat: quiz background와 mobileLayout 분리
